Update Home.jsx
Because of the way the code was formatted, the "Welcome back, {user name}!" doesn't display a space between the comma and the user name (in the latest Chrome is displayed "Welcome back,John Wargo". So I pulled the code in so it spaces properly.
fix lint issue

OKTA-365188
<<<Jenkins Check-In of Tested SHA: 0751d58c966a21520735843cfa1fbba1419a967d for eng_productivity_ci_bot_okta@okta.com>>>
Artifact: samples-js-react